+++ openvkl/drivers/ispc/volume/vdb/VdbIterator.cpp
@@ -18,11 +18,10 @@ namespace openvkl {
         const vrange1fn<W> &tRange,
         const ValueSelector<W> *valueSelector)
     {
-      const auto &vdbVolume = dynamic_cast<const VdbVolume<W> &>(sampler->getVolume());
       CALL_ISPC(VdbIterator_Initialize,
                 static_cast<const int *>(valid),
                 ispcStorage,
-                vdbVolume.getGrid(),
+                sampler->getISPCEquivalent(),
                 (void *)&origin,
                 (void *)&direction,
                 (void *)&tRange,